Tidied for gcc-3.4
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/trunk@15476 72102866-910b-0410-8b05-ffd578937521

+++ Source/Additions/GSXML.m
@@ -3100,7 +3100,8 @@ fatalErrorFunction(void *ctx, const char *msg, ...)
       xmlNode	*node;
       GSXMLNode *n;
 
-      node = xmlXPathNodeSetItem (((xmlXPathObject*)_lib)->nodesetval, index);
+      node = xmlXPathNodeSetItem (((xmlXPathObject*)_lib)->nodesetval,
+	(int)index);
       n = [GSXMLNode alloc];
       
       return [n _initFrom: node  parent: self];
