+++ Source/mframe.m
@@ -358,9 +358,9 @@ mframe_next_arg(const char *typePtr, NSArgumentInfo *info)
 
       case _C_STRUCT_B:
 	{
-	  int acc_size = 0;
-	  int def_align = objc_alignof_type(typePtr-1);
-	  int acc_align = def_align;
+	  unsigned int acc_size = 0;
+	  unsigned int def_align = objc_alignof_type(typePtr-1);
+	  unsigned int acc_align = def_align;
 
 	  /*
 	   *	Skip "<name>=" stuff.
@@ -407,8 +407,8 @@ mframe_next_arg(const char *typePtr, NSArgumentInfo *info)
 
       case _C_UNION_B:
 	{
-	  int	max_size = 0;
-	  int	max_align = 0;
+	  unsigned int	max_size = 0;
+	  unsigned int	max_align = 0;
 
 	  /*
 	   *	Skip "<name>=" stuff.
@@ -1317,7 +1317,7 @@ mframe_build_return (arglist_t argframe,
   /* A pointer to the memory that will hold the return value. */
   retval_t retframe = NULL;
   /* The size, in bytes, of memory pointed to by RETFRAME. */
-  int retsize;
+  unsigned int retsize;
   /* Which argument number are we processing now? */
   int argnum;
   /* Type qualifier flags; see <objc/objc-api.h>. */
@@ -1377,7 +1377,7 @@ mframe_build_return (arglist_t argframe,
 		/* We are returning a pointer to something. */
 		/* Increment TYPE so we can see what it is a pointer to. */
 		tmptype++;
-		retLength = objc_sizeof_type(tmptype);
+		retLength = (unsigned int)objc_sizeof_type(tmptype);
 		/* Allocate memory to hold the value we're pointing to. */
 		*(void**)retframe = 
 		  NSZoneMalloc(NSDefaultMallocZone(), retLength);
